+++ coders/png.c
@@ -3069,7 +3069,7 @@ static Image *ReadPNGImage(const ImageInfo *image_info,ExceptionInfo *exception)
     if (image->delay != 0)
       scenes_found++;
     if (image_info->ping && (image_info->subrange != 0) &&
-        scenes_found > image_info->subimage+image_info->subrange)
+        scenes_found > (long) image_info->subimage+image_info->subrange)
       {
         png_destroy_read_struct(&ping,&ping_info,&end_info);
         break;
@@ -3087,7 +3087,7 @@ static Image *ReadPNGImage(const ImageInfo *image_info,ExceptionInfo *exception)
     png_read_image(ping,scanlines);
     png_read_end(ping,ping_info);
 
-    if (image_info->subrange != 0 && scenes_found-1 < image_info->subimage)
+    if (image_info->subrange != 0 && scenes_found-1 < (long) image_info->subimage)
       {
         png_destroy_read_struct(&ping,&ping_info,&end_info);
         LiberateMemory((void **) &png_pixels);
@@ -4010,7 +4010,7 @@ static Image *ReadPNGImage(const ImageInfo *image_info,ExceptionInfo *exception)
       CatchImageException(image);
       if (image_info->subrange != 0)
         {
-          if (scenes_found > image_info->subimage+image_info->subrange)
+          if (scenes_found > (long) image_info->subimage+image_info->subrange)
             break;
         }
   } while (LocaleCompare(image_info->magick,"MNG") == 0);
